feat(vehicle): 개별 차량 조회 관련 커스터훅 `useSearchSingleVehicle` 구현 (#46) ------------------------------------------------------
------------------------------------------------------
